+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/main/java/com/azure/ai/metricsadvisor/models/AnomalyIncidentStatus.java
@@ -10,6 +10,15 @@ import java.util.Collection;
 
 /** Defines values for AnomalyIncidentStatus. */
 public final class AnomalyIncidentStatus extends ExpandableStringEnum<AnomalyIncidentStatus> {
+
+    /**
+     * Constructs a AnomalyIncidentStatus object.
+     * @deprecated Use the {@link #fromString(String)} factory method.
+     */
+    @Deprecated
+    public AnomalyIncidentStatus() {
+    }
+
     /** Static value Active for AnomalyIncidentStatus. */
     public static final AnomalyIncidentStatus ACTIVE = fromString("Active");
 
@@ -27,7 +36,10 @@ public final class AnomalyIncidentStatus extends ExpandableStringEnum<AnomalyInc
         return fromString(name, AnomalyIncidentStatus.class);
     }
 
-    /** @return known AnomalyIncidentStatus values. */
+    /**
+     * Returns the known AnomalyIncidentStatus values.
+     * @return known AnomalyIncidentStatus values.
+     */
     public static Collection<AnomalyIncidentStatus> values() {
         return values(AnomalyIncidentStatus.class);
     }